STREAMP-3135: add graphql delete functionality integration tests

+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/services/ConsumerBindingService.java
@@ -94,6 +94,11 @@ public class ConsumerBindingService {
 
   @PreAuthorize("hasPermission(#consumerBinding, 'DELETE')")
   public void delete(ConsumerBinding consumerBinding) {
+    val existing = unsecuredGet(consumerBinding.getKey());
+    if (!existing.isPresent()) {
+      throw new ValidationException("Can't delete " + consumerBinding.getKey() + " because it doesn't exist");
+    }
+    handlerService.handleDelete(consumerBinding);
     consumerBindingRepository.delete(consumerBinding);
   }
